Refactor global async for-of known failure tests
