+++ tests/Asynkron.JsEngine.Tests/AsyncIterationTests.cs
@@ -502,100 +502,6 @@ public class AsyncIterationTests(ITestOutputHelper output)
     }
 
     [Fact(Timeout = 2000)]
-    public async Task ForAwaitOf_FallbackToSyncIterator()
-    {
-        var engine = new JsEngine();
-
-        engine.SetGlobalFunction("log", args =>
-        {
-            var message = args.Count > 0 ? args[0]?.ToString() ?? "null" : "null";
-            output.WriteLine($"LOG: {message}");
-            return null;
-        });
-
-        await engine.Run("""
-
-                                     let result = "";
-                                     
-                                     // Object with only sync iterator (Symbol.iterator)
-                                     let syncIterable = {
-                                         [Symbol.iterator]() {
-                                             log("Symbol.iterator called");
-                                             let values = ["x", "y", "z"];
-                                             let index = 0;
-                                             return {
-                                                 next() {
-                                                     log("next() called, index: " + index);
-                                                     if (index < values.length) {
-                                                         let value = values[index++];
-                                                         log("Returning value: " + value);
-                                                         return { value: value, done: false };
-                                                     }
-                                                     log("Done iterating");
-                                                     return { done: true };
-                                                 }
-                                             };
-                                         }
-                                     };
-                                     
-                                     async function test() {
-                                         log("Starting test function");
-                                         log("Starting for-await-of loop");
-                                         for await (let item of syncIterable) {
-                                             log("Got item: " + item);
-                                             result = result + item;
-                                             log("Result so far: " + result);
-                                         }
-                                         log("After loop, result: " + result);
-                                     }
-                                     
-                                     test();
-                                 
-                         """);
-
-        var result = await engine.Evaluate("result;");
-        output.WriteLine($"Final result: '{result}'");
-        Assert.Equal("xyz", result);
-    }
-
-    [Fact(Timeout = 2000)]
-    public async Task ForAwaitOf_WithSyncIteratorNoAsync()
-    {
-        var engine = new JsEngine();
-
-        // Test without async function to isolate the issue
-        var result = await engine.Evaluate("""
-
-                                                       let result = "";
-                                                       
-                                                       // Object with only sync iterator (Symbol.iterator)
-                                                       let syncIterable = {
-                                                           [Symbol.iterator]() {
-                                                               let values = ["x", "y", "z"];
-                                                               let index = 0;
-                                                               return {
-                                                                   next() {
-                                                                       if (index < values.length) {
-                                                                           return { value: values[index++], done: false };
-                                                                       }
-                                                                       return { done: true };
-                                                                   }
-                                                               };
-                                                           }
-                                                       };
-                                                       
-                                                       for await (let item of syncIterable) {
-                                                           result = result + item;
-                                                       }
-                                                       
-                                                       result;
-                                                   
-                                           """);
-
-        Assert.Equal("xyz", result);
-    }
-
-    [Fact(Timeout = 2000)]
     public async Task RegularForOf_WithAwaitInBodyWithDebug()
     {
         // Test that regular for-of with await in body works, using __debug() to inspect state
