+++ tests/Asynkron.JsEngine.Tests/AsyncIteratorDebuggingTests.cs
@@ -737,159 +737,6 @@ public class AsyncIteratorDebuggingTests(ITestOutputHelper output)
     }
 
     [Fact(Timeout = 5000)]
-    public async Task TestF_ActualForAwaitOf_WithLogging()
-    {
-        output.WriteLine("=== Test F: Actual for-await-of with extensive logging ===");
-        
-        var engine = new JsEngine();
-        engine.SetGlobalFunction("log", args =>
-        {
-            var msg = args.Count > 0 ? args[0]?.ToString() ?? "null" : "null";
-            output.WriteLine($"LOG: {msg}");
-            return null;
-        });
-
-        await engine.Run(@"
-            log('Creating global iterable');
-            let globalIterable = {
-                [Symbol.iterator]() {
-                    log('!!! Symbol.iterator called !!!');
-                    let index = 0;
-                    let iterObj = {
-                        next: function() {
-                            log('!!! next() called, index=' + index + ' !!!');
-                            if (index < 2) {
-                                let val = index++;
-                                log('!!! Returning value=' + val + ' !!!');
-                                return { value: val, done: false };
-                            }
-                            log('!!! Returning done=true !!!');
-                            return { done: true };
-                        }
-                    };
-                    log('!!! Returning iterator object !!!');
-                    return iterObj;
-                }
-            };
-            
-            async function test() {
-                log('>>> Starting test function');
-                let result = [];
-                
-                log('>>> About to enter for-await-of');
-                try {
-                    for await (let item of globalIterable) {
-                        log('>>> INSIDE LOOP, item=' + item);
-                        result.push(item);
-                    }
-                    log('>>> After loop, result=' + JSON.stringify(result));
-                } catch (e) {
-                    log('>>> Exception in loop: ' + e);
-                    throw e;
-                }
-                
-                log('>>> Returning result');
-                return result;
-            }
-            
-            log('Calling test()');
-            test()
-                .then(r => log('FINAL: ' + JSON.stringify(r)))
-                .catch(e => log('ERROR: ' + e));
-        ");
-
-        await System.Threading.Tasks.Task.Delay(1000);
-        output.WriteLine("Actual for-await-of test with extensive logging to pinpoint failure");
-    }
-
-    [Fact(Timeout = 5000)]
-    public async Task TestG_CaptureExceptionsWithChannel()
-    {
-        output.WriteLine("=== Test G: Capture Exceptions via Exception Channel ===");
-        
-        var engine = new JsEngine();
-        engine.SetGlobalFunction("log", args =>
-        {
-            var msg = args.Count > 0 ? args[0]?.ToString() ?? "null" : "null";
-            output.WriteLine($"LOG: {msg}");
-            return null;
-        });
-
-        await engine.Run(@"
-            log('Creating global iterable with iterator');
-            let globalIterable = {
-                [Symbol.iterator]() {
-                    log('Symbol.iterator called');
-                    let index = 0;
-                    return {
-                        next: function() {
-                            log('next() called, index=' + index);
-                            if (index < 2) {
-                                return { value: index++, done: false };
-                            }
-                            return { done: true };
-                        }
-                    };
-                }
-            };
-            
-            async function test() {
-                log('Starting test');
-                let result = [];
-                
-                try {
-                    log('Entering for-await-of');
-                    for await (let item of globalIterable) {
-                        log('Got item: ' + item);
-                        result.push(item);
-                    }
-                    log('Loop completed, result: ' + JSON.stringify(result));
-                } catch (e) {
-                    log('Caught exception: ' + e);
-                }
-                
-                return result;
-            }
-            
-            test().then(r => log('Done: ' + JSON.stringify(r)))
-                .catch(e => log('Failed: ' + e));
-        ");
-
-        await System.Threading.Tasks.Task.Delay(1000);
-
-        // Check for exceptions in the exception channel
-        var exceptions = new List<ExceptionInfo>();
-        while (engine.Exceptions().TryRead(out var ex))
-        {
-            exceptions.Add(ex);
-        }
-
-        output.WriteLine("");
-        output.WriteLine($"=== EXCEPTIONS CAPTURED: {exceptions.Count} ===");
-        foreach (var ex in exceptions)
-        {
-            output.WriteLine($"Exception: {ex.ExceptionType}");
-            output.WriteLine($"Message: {ex.Message}");
-            output.WriteLine($"Context: {ex.Context}");
-            output.WriteLine($"Call Stack:");
-            foreach (var frame in ex.CallStack)
-            {
-                output.WriteLine($"  - {frame.Description}");
-            }
-            output.WriteLine("");
-        }
-
-        if (exceptions.Count > 0)
-        {
-            output.WriteLine("✅ SUCCESS: Captured exceptions that explain the failure!");
-        }
-        else
-        {
-            output.WriteLine("⚠️ No exceptions captured - exception may be swallowed elsewhere");
-        }
-    }
-
-    [Fact(Timeout = 5000)]
     public async Task TestH_CheckPromiseRejectionHandling()
     {
         output.WriteLine("=== Test H: Check if Promise Rejections are Handled ===");
