+++ tests/Asynkron.JsEngine.Tests/MethodShorthandParsingTests.cs
@@ -184,72 +184,4 @@ public class MethodShorthandParsingTests(ITestOutputHelper output)
         output.WriteLine("This tests method shorthand in progressively complex scenarios");
     }
 
-    [Fact(Timeout = 5000)]
-    public async Task TestL_MethodShorthandInForAwaitOf()
-    {
-        output.WriteLine("=== Test L: Method Shorthand in for-await-of ===");
-        
-        var engine = new JsEngine();
-        engine.SetGlobalFunction("log", args =>
-        {
-            var msg = args.Count > 0 ? args[0]?.ToString() ?? "null" : "null";
-            output.WriteLine($"LOG: {msg}");
-            return null;
-        });
-
-        await engine.Run(@"
-            log('Creating global iterable with method shorthand next()');
-            let globalIterable = {
-                [Symbol.iterator]() {
-                    log('Symbol.iterator() called');
-                    let count = 0;
-                    return {
-                        next() {
-                            log('next() called, count=' + count);
-                            if (count < 2) {
-                                return { value: count++, done: false };
-                            }
-                            return { done: true };
-                        }
-                    };
-                }
-            };
-            
-            async function testAsync() {
-                log('Starting async function');
-                let results = [];
-                log('Entering for-await-of loop');
-                for await (let item of globalIterable) {
-                    log('Loop body: got item = ' + item);
-                    results.push(item);
-                }
-                log('Exited loop, results: ' + JSON.stringify(results));
-                return results;
-            }
-            
-            log('Calling testAsync()');
-            testAsync()
-                .then(r => log('SUCCESS: ' + JSON.stringify(r)))
-                .catch(e => log('ERROR: ' + e));
-        ");
-
-        await System.Threading.Tasks.Task.Delay(1000);
-
-        // Check exceptions
-        var exceptions = new List<ExceptionInfo>();
-        while (engine.Exceptions().TryRead(out var ex))
-        {
-            exceptions.Add(ex);
-        }
-
-        output.WriteLine("");
-        output.WriteLine($"=== EXCEPTIONS: {exceptions.Count} ===");
-        foreach (var ex in exceptions)
-        {
-            output.WriteLine($"  - {ex.Message} (Context: {ex.Context})");
-        }
-
-        output.WriteLine("");
-        output.WriteLine("This is the EXACT scenario from the failing test!");
-    }
 }
